pyproject.toml
[tool.poetry]
name = "streamscharts-top100-reports"
version = "0.1.0"
description = ""
authors = ["Simon Roadknight <s.roadknight96@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
duckdb = "^1.1.3"
pandas = "^2.2.3"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
